refactor bounding box code into its own file
