+++ src/js/utils/math.ts
@@ -1,5 +1,4 @@
 import { Keypoint } from "@tensorflow-models/pose-detection";
-import { min, max } from "lodash";
 
 /**
  * uses the Pythagorean theorem to calculate the distance between keypoints in 2D, or 3D (if available)
@@ -61,22 +60,4 @@ export const calculateLineWidth = (keypoints: Keypoint[]) => {
     LINE_WIDTH_MIN,
     LINE_WIDTH_MAX
   );
-};
-
-export const getBoundingBox = (keypoints: Keypoint[]) => {
-  const xS = keypoints.map((keypoint) => keypoint.x);
-  const yS = keypoints.map((keypoint) => keypoint.y);
-  const minX = min(xS) as number;
-  const minY = min(yS) as number;
-  const maxX = max(xS) as number;
-  const maxY = max(yS) as number;
-
-  return {
-    minX,
-    minY,
-    maxX,
-    maxY,
-    width: maxX - minX,
-    height: maxY - minY,
-  };
 };
